api/app/requirements.txt
Django==5.2
djangorestframework
drf_yasg
django-cors-headers
djangorestframework-simplejwt